fix: output emissions in fixed and predicted order
emissions was previously outputted in arbitrary order, that made it often difficult to compare output of a model that had been changed, or run with different versions of eCalc. Now the order should be the same, and since the name of an emissions is now a key, it will be easier to correctly compare different versions of output.