ci: avoid conflicting releases
To allow both for easy patching of a release on a separate
release branch we need to create a release branch that we can
patch and only bump patch versions to when needed.

The patch bump should therefore be reserved for that branch to
avoid that a new release on main branch will get a conflicting
version. Main will therefore in general only have minor bumps,
and a release to release should be a minor bump.

If we want to have a major bump, due to breaking changes or
many changes and improvements since a previous major version
we should do a manual trigger of release-please with
that specified as an argument.