chore: remove limiting dependency typer-cli
We use typer-cli to auto-generate the CLI reference
for the eCalc CLI. However typer-cli is rarely upgraded
and is currently limited to use typer 0.7.0 while typer
is at 0.9.0. This restricts us unnecessarily to use the latest
typer and is potentially a security concern.

The small portion of the code that we were using in typer-cli
have been extracted together with the license details to still
be able to use the functionality and avoid the limitation
of the typer dependency.