chore: simplify dependencies for use with komodo
typer optionally depends on shellingham, rich and
colorama. Rich is already an explicit dependency in ecalc,
while colorama is a dependency through click and
already in komodo.

By only depending on typer and not typer[all] we remove
the need for shellingham, as that dependency should not
be needed in komodo. It is used to install auto-completion
for ecalc, and it automatically detects the shell. If the
user still wants auto-completion, they will need to specify
shell explicitly with the argument.

We should aim to keep the number of dependencies to a minimum,
in general, but in particular with Komodo, in order to make
it easier to maintain.