chore: merge queque
* chore: remove unnecessary noqa statements

* chore: sys-exit-alias (#114)

No guarantee that the exit() from site module is available at runtime. Prefer sys.exit(), as the sys module is guaranteed to exist in all contexts.

* refactor: reduce branching in if-statements (#115)

* chore: sys-exit-alias

No guarantee that the exit() from site module is available at runtime. Prefer sys.exit(), as the sys module is guaranteed to exist in all contexts.

* refactor: reduce branching in if-statements

* refactor: reduce branching in for-loop