fix: make ecalc installable again
When restructuring libecalc the previous mappings
for which packages was removed for simplicity. We
need to include those in order to make poetry/pip
include the required packages in src/ directory, since
it will only include libecalc by default (it will look
for src dir and package name). It is nevertheless better
to be explicit.

The current functionality is the same way as it worked
before it failed.