fix: avoid name conflicts with ecalc cli package
Using the generic name CLI and including that as a part of
libecalc, installs it as a separate package in the venv called
"cli". This is too general/generic to use, as it will easily
cause a name collision with another package, also it does not
give any information of which package it depends to (libecalc).