chore: clarify neqsim depenedency in ecalc (#198)
To more easily see connection and association with ecalc, it
is better to have ecalc as a prefix in package name of the
neqsim-wrapper when being installed. We still want to keep the
wrapper on the "outside" of the library, to make it more clear
that it should be pluggable and optional.