refactor: remove regularity our of core/domain
Regularity should not be transferred or handled in
core. At the time data arrives in core, it should only
be stream day, and we should not have to relate to "what
type of rate" is it, until we decide whether the user
output should be calendar or stream day (default) based
on user needs.

Refs ECALC-196