refactor: move into presentation layer (#271)
Our code in libeCalc is messy, and we need a better way to organize it. We also need to be more strict wrt which parts of the code that can call and depend on other parts, due to both cyclic dependencies and to make it easier to change the code.

Currently I only moved things from the input and output module to the presentation module. A few things were added to infrastructure, since it has to deal with filesystem. Both input and output has to do with presentation and "communication" with users and systems, and is often two sides of the same thing (request=input, respons=output), but we happen to have yaml as input and json as output, through different use cases.

For now, it is just structured according to presentation, infrastructure, application and domain. This may change later, if we think that is better. However, the structuring in different modules such as "flow diagram" and "query export", which are both "applications" and use cases of our library, should later be more modularized and separated, but for a small step to make things a little better, this was the judgement I made now...

TLDR; no code changes, just moving code....so if done correctly that should not affect git history

NB: I also had to run pre-commit run --all due to alphabetical changes in the imports, to make the CI happy.
Also, I had to ignore in api documentation, as something seems to have caused cyclic references [sic]