chore: changing algorithm for get max rate for streams
The get max standard rate algorithm for a multiple streams and
pressures compressor train does not work fully as intended.
Changing to a brute force algorithm, instead of trying to be clever
in searching along max speed curve and stone wall etc.

This algorithm will of course be more computationally demainding.