feat: allow cross overs both in and out of the same compressor train in a compressor system
This can be done if it is made sure that the compressor trains in the compressor system
are processed in the right order. You can only pass rates to a compressor train that
has not been processed yet. Hence, the indexes in the list specifying the cross overs
must be larger than the index of the compressor train where the cross over comes from
(or zero, meaning no cross over).