fix: don't raise error when missing time series on token collection
When collecting the tokens from time series collection, we don't want to
raise an error since that would limit the information we get from
validation. Instead we want to collect the tokens we can, then validate
the model based on that context. The errors will still give information
about a missing resource, but also additional information, and file
context.