fix: issue with empty optional keywords
If specifying only a keyword without the data for an optional keyword in
the yaml we would not get a validation error. I.e. the json-schema did
not capture the error. In addition, the validation context setup did not
account for None in the data, expecting either unset or some data (not
None). This caused a type error since we then tried to use None as a
list.

Adding only `MODELS:` to the yaml would reproduce the error. Removing
`MODELS:` or adding an empty list `MODELS: []` would make it valid.