fix: some models crashes with monthly frequency (#1059)
Turns out that max shore calculation for LTP had not been updated
to use resampling for values, but only for periods. There was
therefore a mismatch when running LTP when data-defined periods
happened to not match monthly frequency resolution.

Bonus bug fix as well, that makes it possible to run .resample when we 
use data-defined frequency, as it will return the passed in model without
any resampling.

Refs: equinor/ecalc-internal#1028