fix: also supply gasTurbineGeneratorConsumption for ltp if empty
If we have no matching category and query filter for any periods
for a given ecalc model, the gasTurbineGeneratorConsumption column
was skipped. It will now always be included.

This is needed in order to have a consistent LTP format even when
columns are empty, for comparison and for export to centuries.

Refs: equinor/ecalc-internal#1031