feat: experimental introduction of alternative neqsim impl (#1120)
We are changing from the legacy Py4J implementation of
Neqsim, where we bundle the neqsim.jar as a part of this 
repo, to the JPype implementation of Neqsim provided by
the Neqsim team. They also bundle the jar through the
jneqsim dependency, so that neqsim python is up to date 
with java neqsim. Additionally, jpype should be faster and
easier to use than py4j.

To allow for a smooth change and be able to test JPype
and JNeqsim safely, we introduce the change via a
command line argument, or a feature flag to be able to
choose which implementation to use, on startup.

I have decided to use a "singleton factory", where 
the developer has to set which implementation to
instantiate upon initialization. In order to change, a 
restart of the application is needed.

Refs: equinor/ecalc-internal#1068