refactor: add EnergyResult definition
composition instead of inheritance for energy results to separate it
from process results, i.e. CompressorTrainResult has EnergyResult
instead of being an energy result. Providing EnergyResult on-demand also
gives us flexibility in terms of when we calculate the result, i.e. we
can show clearly that power for CompressorTrainResult is the sum of
power for compressor stages by only keeping stage results and
calculating the sum when asked for.