refactor: add ChartData interface
The ChartData interface allows different implementation of charts, which
makes it possible for us to handle generic from input in the same
interface. ChartData allows for delayed calculation of curves, which is
necessary to avoid heavy computations on instantiation of
GenericFromInput charts. GenericFromInput will be implemented as
ChartData in a separate PR.